chore: move lookup view to dendron-plugin-views
